WARNING READ BEFORE GOING TO THE MAIN SITE index.html
WARNING: THIS BETA VERSION MIGHT DELETE ALL YOUR DATA, DONT ACCESS THE MAIN SITE RIGHT NOW. as the developer of this game i do not trust my own beta versions, they are called BETA for a reason because they are unfinished versions of the game, this version changed the entire code of the game which means old data may be deleted. Because this version introduced different accounts which you can access them, however this is not the case because the main problem is the .windowonload and localstorage aka your saved data might be deleted of this new update.